Refactor IstioDetails Overview Tab
